+++ NEWS.md
@@ -33,6 +33,12 @@
   invalid codec ID (e.g. `V_MPEG7`) in certain cases. Fixes #1995.
 * mkvmerge: MPEG PS reader: made the file type detection less strict so that
   garbage at the start of the file doesn't prevent detection. Fixes #2008.
+* mkvmerge: JSON identification: the `stream_id` and `sub_stream_id` track
+  properties were output as hexadecimal strings instead of unsigned
+  integers. As the `ts_pid` track property was only used for MPEG transport
+  streams, its value is now output as `stream_id` instead, and the `ts_pid`
+  property has been removed. The JSON schema version has been bumped to 8 due
+  to this change.
 
 
 # Version 12.0.0 "Trust / Lust" 2017-05-20